Refactor common functionality into util
One flat module for the limited set of common functions yet.
