+++ src/fmu/ensemble/realization.py
@@ -17,7 +17,6 @@ import copy
 import glob
 import json
 from datetime import datetime, date, time
-import calendar
 import dateutil
 
 import yaml
@@ -32,6 +31,13 @@ from ecl import EclFileFlagEnum
 from .etc import Interaction
 from .virtualrealization import VirtualRealization
 from .realizationcombination import RealizationCombination
+from .util import (
+    parse_number,
+    unionize_smry_dates,
+    flatten,
+    compute_volumetric_rates,
+    shortcut2path,
+)
 
 HAVE_ECL2DF = False
 try:
@@ -681,10 +687,10 @@ class ScratchRealization(object):
             KeyError if data is not found.
             TypeError if data in localpath or merge is not of a mergeable type
         """
-        fullpath = self.shortcut2path(localpath)
+        fullpath = shortcut2path(self.keys(), localpath)
         if fullpath not in self.data.keys():
             raise KeyError("Could not find {}".format(localpath))
-        data = self.data[self.shortcut2path(localpath)]
+        data = self.data[shortcut2path(self.keys(), localpath)]
         if not isinstance(merge, list):
             merge = [merge]  # can still be None
         if merge and merge[0] is not None:
@@ -727,45 +733,6 @@ class ScratchRealization(object):
                     + "from {} of type {}".format(mergekey, type(data))
                 )
         return data
-
-    def shortcut2path(self, shortpath):
-        """
-        Convert short pathnames to fully qualified pathnames
-        within the datastore.
-
-        If the fully qualified localpath is
-
-          'share/results/volumes/simulator_volume_fipnum.csv'
-
-        then you can also access this with these alternatives:
-
-          * simulator_volume_fipnum
-          * simulator_volume_fipnum.csv
-          * share/results/volumes/simulator_volume_fipnum
-
-        but only as long as there is no ambiguity. In case
-        of ambiguity, the shortpath will be returned as is.
-        """
-        basenames = list(map(os.path.basename, self.data.keys()))
-        if basenames.count(shortpath) == 1:
-            short2path = {os.path.basename(x): x for x in self.data}
-            return short2path[shortpath]
-        noexts = ["".join(x.split(".")[:-1]) for x in self.data]
-        if noexts.count(shortpath) == 1:
-            short2path = {"".join(x.split(".")[:-1]): x for x in self.data}
-            return short2path[shortpath]
-        basenamenoexts = [
-            "".join(os.path.basename(x).split(".")[:-1]) for x in self.data
-        ]
-        if basenamenoexts.count(shortpath) == 1:
-            short2path = {
-                "".join(os.path.basename(x).split(".")[:-1]): x for x in self.data
-            }
-            return short2path[shortpath]
-        # If we get here, we did not find anything that
-        # this shorthand could point to. Return as is, and let the
-        # calling function handle further errors.
-        return shortpath
 
     def find_files(self, paths, metadata=None, metayaml=False):
         """Discover realization files. The files dataframe
@@ -1217,154 +1184,9 @@ class ScratchRealization(object):
     def get_volumetric_rates(self, column_keys=None, time_index=None, time_unit=None):
         """Compute volumetric rates from cumulative summary vectors
 
-        Column names that are not referring to cumulative summary
-        vectors are silently ignored.
-
-        A Dataframe is returned with volumetric rates, that is rate
-        values that can be summed up to the cumulative version. The
-        'T' in the column name is switched with 'R'. If you ask for
-        FOPT, you will get FOPR in the returned dataframe.
-
-        Rates in the returned dataframe are valid **forwards** in time,
-        opposed to rates coming directly from the Eclipse simulator which
-        are valid backwards in time.
-
-        If time_unit is set, the rates will be scaled to represent
-        either daily, monthly or yearly rates. These will sum up to the
-        cumulative as long as you multiply with the correct number
-        of days, months or year between each consecutive date index.
-        Month lengths and leap years are correctly handled.
-
-        The returned dataframe is indexed by DATE.
-
-        Args:
-            column_keys: str or list of strings, cumulative summary vectors
-            time_index: str or list of datetimes
-            time_unit: str or None. If None, the rates returned will
-                be the difference in cumulative between each included
-                time step (where the time interval can vary arbitrarily)
-                If set to 'days', 'months' or 'years', the rates will
-                be scaled to represent a daily, monthly or yearly rate that
-                is compatible with the date index and the cumulative data.
-
+        See :meth:`fmu.ensemble.util.compute_volumetric_rates`
         """
-        return self.static_get_volumetric_rates(
-            self, column_keys, time_index, time_unit
-        )
-
-    @staticmethod
-    def static_get_volumetric_rates(realization, column_keys, time_index, time_unit):
-        """Static method for volumetric rates
-
-        This method is to be used by both ScratchRealization
-        and VirtualRealization, and is documented there."""
-
-        if isinstance(time_unit, str):
-            if time_unit not in ["days", "months", "years"]:
-                raise ValueError(
-                    "Unsupported time_unit " + time_unit + " for volumetric rates"
-                )
-
-        column_keys = realization._glob_smry_keys(column_keys)
-
-        # Be strict and only include certain summary vectors that look
-        # cumulative by their name:
-        column_keys = [
-            x for x in column_keys if ScratchRealization._cum_smrycol2rate(x)
-        ]
-        if not column_keys:
-            logger.error("No valid cumulative columns given to volumetric computation")
-            return pd.DataFrame()
-
-        cum_df = realization.get_smry(column_keys=column_keys, time_index=time_index)
-        # get_smry() for realizations return a dataframe indexed by 'DATE'
-
-        # Compute row-wise difference, shift back one row
-        # to get the NaN to the end, and then drop the NaN.
-        # The "rate" given for a specific date is then
-        # valid from that date until the next date.
-        diff_cum = cum_df.diff().shift(-1).fillna(value=0)
-
-        if time_unit:
-            # Calculate the relative timedelta between consecutive
-            # DateIndices. relativedeltas are correct in terms
-            # of number of years and number of months, but it will
-            # only give us integer months, and then leftover days.
-            rel_deltas = [
-                dateutil.relativedelta.relativedelta(t[1], t[0])
-                for t in zip(diff_cum.index, diff_cum.index[1:])
-            ]
-            whole_days = [
-                (t[1] - t[0]).days for t in zip(diff_cum.index, diff_cum.index[1:])
-            ]
-            # Need to know which years are leap years for our index:
-            dayspryear = [
-                365 if not calendar.isleap(x.year) else 366
-                for x in pd.to_datetime(diff_cum.index[1:])
-            ]
-            # Float-contribution to years from days:
-            days = [
-                t[0] / float(t[1])
-                for t in zip([r.days for r in rel_deltas], dayspryear)
-            ]
-            floatyearsnodays = [r.years + r.months / 12.0 for r in rel_deltas]
-            floatyears = [x + y for x, y in zip(floatyearsnodays, days)]
-
-            # Calculate month-difference:
-            floatmonthsnodays = [r.years * 12.0 + r.months for r in rel_deltas]
-            # How many days pr. month? We check this for the right
-            # end of the relevant time interval.
-            daysprmonth = [
-                calendar.monthrange(t.year, t.month)[1] for t in diff_cum.index[1:]
-            ]
-            days = [
-                t[0] / float(t[1])
-                for t in zip([r.days for r in rel_deltas], daysprmonth)
-            ]
-            floatmonths = [x + y for x, y in zip(floatmonthsnodays, days)]
-
-            diff_cum["DAYS"] = whole_days + [0]
-            diff_cum["MONTHS"] = floatmonths + [0]
-            diff_cum["YEARS"] = floatyears + [0]
-            for vec in column_keys:
-                diff_cum[vec] = diff_cum[vec] / diff_cum[time_unit.upper()]
-            # Drop temporary columns
-            diff_cum.drop(["DAYS", "MONTHS", "YEARS"], inplace=True, axis=1)
-            # Set NaN at the final row to zero
-            diff_cum.fillna(value=0, inplace=True)
-
-        # Translate the column vectors, 'FOPT' -> 'FOPR' etc.
-        rate_names = []
-        for vec in diff_cum.columns:
-            ratename = ScratchRealization._cum_smrycol2rate(vec)
-            if ratename:
-                rate_names.append(ratename)
-        diff_cum.columns = rate_names
-        diff_cum.index.name = "DATE"
-        return diff_cum
-
-    @staticmethod
-    def _cum_smrycol2rate(smrycolumn):
-        """Returns None if a smrycolumn is not assumed
-        to be cumulative, and returns a string with the corresponding
-        rate column if it is cumulative
-
-        F.ex. _cum_smrycol2rate('FOPT') will return 'FOPR'
-        """
-        # Split by colon into components:
-        comps = smrycolumn.split(":")
-        if len(comps) > 2:
-            # Do not support more than one colon.
-            return None
-        if "CT" in comps[0]:
-            # No watercuts.
-            return None
-        if "T" not in comps[0]:
-            return None
-        comps[0] = comps[0].replace("T", "R")
-        if len(comps) > 1:
-            return comps[0] + ":" + comps[1]
-        return comps[0]
+        return compute_volumetric_rates(self, column_keys, time_index, time_unit)
 
     def get_smryvalues(self, props_wildcard=None):
         """
@@ -1432,13 +1254,10 @@ class ScratchRealization(object):
         Returns:
             list of datetimes. None if no summary data is available.
         """
-        # pylint: disable=import-outside-toplevel
-        from .ensemble import ScratchEnsemble
-
         eclsum = self.get_eclsum(include_restart=include_restart)
         if not eclsum:
             return None
-        return ScratchEnsemble._get_smry_dates(
+        return unionize_smry_dates(
             [eclsum.dates], freq, normalize, start_date, end_date
         )
 
@@ -1466,7 +1285,7 @@ class ScratchRealization(object):
             criteria.
         """
         kwargs.pop("inplace", 0)
-        localpath = self.shortcut2path(localpath)
+        localpath = shortcut2path(self.keys(), localpath)
         if localpath not in self.keys():
             return False
         if not kwargs:
@@ -1527,7 +1346,7 @@ class ScratchRealization(object):
                 dataframes
             keys: list of strings of keys to delete from a dictionary
         """
-        fullpath = self.shortcut2path(localpath)
+        fullpath = shortcut2path(self.keys(), localpath)
         if fullpath not in self.keys():
             raise ValueError("%s not found" % localpath)
 
@@ -1746,103 +1565,3 @@ class ScratchRealization(object):
             non-active cells are given value 0.
         """
         return self.get_unrst()[prop][report].scatter_copy(self.actnum)
-
-
-def normalize_dates(start_date, end_date, freq):
-    """
-    Normalize start and end date according to frequency
-    by extending the time range.
-
-    So for [1997-11-5, 2020-03-02] and monthly freqency
-    this will transform your dates to
-    [1997-11-1, 2020-04-01]
-
-    For yearly frequency will be [1997-01-01, 2021-01-01].
-
-    Args:
-        start_date: datetime.date
-        end_date: datetime.date
-        freq: string with either 'monthly' or 'yearly'.
-            Anything else will return the input as is
-    Return:
-        Tuple of normalized (start_date, end_date)
-    """
-
-    if freq == "monthly":
-        start_date = start_date.replace(day=1)
-
-        # Avoid rolling forward if we are already at day 1 in a month
-        if end_date != end_date.replace(day=1):
-            end_date = end_date.replace(day=1) + dateutil.relativedelta.relativedelta(
-                months=1
-            )
-    elif freq == "yearly":
-        start_date = start_date.replace(day=1, month=1)
-        # Avoid rolling forward if we are already at day 1 in a year
-        if end_date != end_date.replace(day=1, month=1):
-            end_date = end_date.replace(
-                day=1, month=1
-            ) + dateutil.relativedelta.relativedelta(years=1)
-    elif freq == "daily":
-        # This we don't need to normalize, but we should not give any warnings
-        pass
-    elif freq == "first" or freq == "last":
-        # This we don't need to normalize, but we should not give any warnings
-        pass
-    else:
-        logger.warning("Unrecognized frequency %s for date normalization", str(freq))
-    return (start_date, end_date)
-
-
-def flatten(dictionary, parent_key="", sep="_"):
-    """Flatten dictionary keys
-
-    Code from stackoverflow.
-    """
-    # pylint: disable=import-outside-toplevel
-    try:
-        # Python3.3^
-        from collections.abc import MutableMapping
-    except ImportError:
-        # Python 2.7:
-        from collections import MutableMapping
-    items = []
-    for key, value in dictionary.items():
-        new_key = parent_key + sep + key if parent_key else key
-        if isinstance(value, MutableMapping):
-            items.extend(flatten(value, new_key, sep=sep).items())
-        else:
-            items.append((new_key, value))
-    return dict(items)
-
-
-def parse_number(value):
-    """Try to parse the string first as an integer, then as float,
-    if both fails, return the original string.
-
-    Caveats: Know your Python numbers:
-    https://stackoverflow.com/questions/379906/how-do-i-parse-a-string-to-a-float-or-int-in-python
-
-    Beware, this is a minefield.
-
-    Returns:
-        int, float or string
-    """
-    if isinstance(value, int):
-        return value
-    if isinstance(value, float):
-        # int(afloat) fails on some, e.g. NaN
-        try:
-            if int(value) == value:
-                return int(value)
-            return value
-        except ValueError:
-            return value  # return float
-    else:  # noqa
-        try:
-            return int(value)
-        except ValueError:
-            try:
-                return float(value)
-            except ValueError:
-                return value
